contracts/pyproject.toml
[tool.poetry]
name = "defacto-contracts"
version = "0.1.0"
description = "DeFacto Protocol Smart Contracts for Algorand"
authors = ["DeFacto Team"]

[tool.poetry.dependencies]
python = "^3.11"
algorand-python = "^1.0.0"
py-algorand-sdk = "^2.5.0"
pyteal = "^0.25.0"
beaker-pyteal = "^1.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
mypy = "^1.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.algokit]
min_version = "2.0.0"

[tool.algokit.project]
type = "contract"
name = "defacto-contracts"
artifacts = "./artifacts"